Refactor Scene and Renderer for Improved Scene Management
- Introduced SceneOffset struct to handle translation of child scenes when appending.
- Replaced merge functionality with append method to allow positioning of child scenes.
- Updated Scene methods to accommodate new append logic, ensuring proper translation of geometry and brushes.
- Enhanced tests to validate new append behavior, including position checks and clipping.
- Adjusted various components to ensure compatibility with new translation logic, including text glyphs and draw records.
- Fixed translation of light sources and filters during scene appending.
- Updated SVG handling to utilize the new append method for scene integration.

+++ src/shared/offscreen.rs
@@ -183,7 +183,7 @@ fn translated_scene_for_bounds(scene: &Scene, local: LocalSpace) -> Scene {
         .text_glyphs
         .iter()
         .copied()
-        .map(|glyph| glyph.translated(-local.surface.x0, -local.surface.y0))
+        .map(|glyph| glyph.translated(-f64::from(local.surface.x0), -f64::from(local.surface.y0)))
         .collect();
     translated.text_runs = scene.text_runs.clone();
     translated.bd_records = translated_backdrop_records(scene, &translated);
